New code outline
