+++ package.json
@@ -33,6 +33,7 @@
     "tslib": "^1.11.1"
   },
   "devDependencies": {
+    "@types/debug": "^4.1.5",
     "@types/fluent-ffmpeg": "^2.1.14",
     "@types/fs-extra": "^8.1.0",
     "@types/jest": "^25.1.3",
