+++ src/VideoWriter.ts
@@ -1,0 +1,3 @@
+export class VideoWriter {
+  // create ffmpeg
+}
