+++ src/types.ts
@@ -1,0 +1,5 @@
+export interface ScreencastFrame {
+  data: Buffer;
+  received: number;
+  timestamp: number;
+}
